# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Book Management API",
        "description": "The API for the Book Management app",
    "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
    }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
        {
            "name": "Books",
            "description": "API for books"
        },
        {
            "name": "Token",
            "description": "API for Authorization"
        }
    ],
    "paths": {
        "/token":{
            "get":{
                "tags":["Token"],
                "summary": "REST endpoint to get a token. At this moment the REST endpoint provide no token and just sends a 200, because the Authorization is not implemented.",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "501": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/books": {
            "get":{
                "tags":["Books"],
                "summary": "REST endpoint to get all books.",
                "responses": {
                    "200": {
                        "description": "A JSON array of book objects.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": "string",
                                            "title": "string",
                                            "author": "string",
                                            "publisher": "string",
                                            "publishYear": "number",
                                            "borrowed": "boolean",
                                            "borrowedDate": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "There are no books on the server."
                    },
                    "501":{
                        "description": "Internal server error."
                    }
                }
            },
            "post":{
                "tags":["Books"],
                "summary": "REST endpoint to create a new book.",
                "requestBody":{
                    "required": "true",
                    "description": "The data to create a new book.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "id": "string",
                                        "title": "string",
                                        "author": "string",
                                        "publisher": "string",
                                        "publishYear": "number",
                                        "borrowed": "boolean",
                                        "borrowedDate": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The book was created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": "string",
                                            "title": "string",
                                            "author": "string",
                                            "publisher": "string",
                                            "publishYear": "number",
                                            "borrowed": "boolean",
                                            "borrowedDate": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Internal server error."
                    }
                }
            },
            "put":{
                "tags":["Books"],
                "summary": "Update a book.",
                "parameters": [
                    {
                        "name": "id",
                        "required": "true",
                        "schema": {
                            "type": "string" 
                        },
                        "in": "path",
                        "description": "The id of the book."
                    }
                ],
                "responses":{
                    "200": {
                        "description": "The book was updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": "string",
                                            "title": "string",
                                            "author": "string",
                                            "publisher": "string",
                                            "publishYear": "number",
                                            "borrowed": "boolean",
                                            "borrowedDate": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Internal server error."
                    }
                }
            },
            "delete":{
                "tags":["Books"],
                "summary": "Delete a book.",
                "parameters": [
                    {
                        "name": "id",
                        "required": "true",
                        "schema": {
                            "type": "string" 
                        },
                        "in": "path",
                        "description": "The id of the book."
                    }
                ],
                "responses":{
                    "200": {
                        "description": "The book was deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": "string",
                                            "title": "string",
                                            "author": "string",
                                            "publisher": "string",
                                            "publishYear": "number",
                                            "borrowed": "boolean",
                                            "borrowedDate": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Internal server error."
                    }
                }
            }
        }
    }
}